refactor: externalize GridRenderBehavior to a behavior
